[e-acsl] add binary operators (&&) and (||) and unary operators (!) and (-) in interlang

Closes e-acsl#307

See merge request frama-c/frama-c!5762



Approved-by: Jordan ISCHARD <jordan.ischard@cea.fr>

+++ src/plugins/e-acsl/tests/bts/oracle/gen_bts2231.c
@@ -36,7 +36,7 @@ int main(void)
     __gmpz_add(__gen_e_acsl_add,
                (__e_acsl_mpz_struct const *)(__gen_e_acsl_A),
                (__e_acsl_mpz_struct const *)(__gen_e_acsl__2));
-    __gmpz_init_set_si(__gen_e_acsl__3,(long)(-1));
+    __gmpz_init_set_si(__gen_e_acsl__3,-1L);
     __gen_e_acsl_eq = __gmpz_cmp((__e_acsl_mpz_struct const *)(__gen_e_acsl_add),
                                  (__e_acsl_mpz_struct const *)(__gen_e_acsl__3));
     __e_acsl_assert_register_long(& __gen_e_acsl_assert_data,"A",0,A);
